Add UI components and update manager for Helldivers Numpad Macros
- Introduced `ui_components.py` containing reusable UI widgets and dialogs for macro testing and settings management.
- Implemented `TestEnvironment`, `SettingsDialog`, and `SettingsWindow` classes for user interaction.
- Added `DraggableIcon` and `NumpadSlot` for managing stratagems visually.
- Created `update_manager.py` to handle update checking, downloading, and installation dialogs.
- Integrated download functionality with progress indication and error handling.
- Added dialogs for notifying users of available updates and managing installations.
